Split functions.go into call.go
